tests/wpt/tests/css/css-align/content-distribution/place-content-shorthand-005.html

<!DOCTYPE html>
<title>CSS Box Alignment: place-content shorthand - inherit value</title>
<link rel="author" title="Javier Fernandez" href="mailto:jfernandez@igalia.com" />
<link rel="help" href="http://www.w3.org/TR/css3-align/#propdef-place-content" />
<meta name="assert" content="Check that place-content's 'inherit' value expands to 'align-content' and 'justify-content'." />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    #test {
        place-content: start end;
    }
</style>
<div id="log"></div>
<div id="test"></div>
<script>
    var child = document.createElement("div");
    document.getElementById("test").appendChild(child);
    child.setAttribute("style", "place-content: inherit");
    var style = getComputedStyle(child);

    test(function() {
        assert_equals(style.getPropertyValue("align-content"),
            "start", "place-content resolved value for align-content");
    }, "Check place-content: inherit - align-content resolved value");

    test(function() {
        assert_equals(style.getPropertyValue("justify-content"),
           "end", "place-content resolved value for justify-content");
    }, "Check place-content: inherit - justify-content resolved value");
</script>
